nutrition-analyzer-flash/requirements.txt
Pillow
pandas
python-dotenv
google-generativeai
clarifai-grpc